requirements.txt
cs50
Flask
Flask-Session
Flask[async]
requests
deepl
deepgram-sdk
werkzeug.security
functools